Move preferences into focused file
